Fix request tracker cookie delete (#549)
Currently when stopping tracking a request the SP tried to delete the relevant cookie by setting it again with an empty value and and expired time.
This doesn't work since the path doesn't match the one of the original cookie.
Fixed by setting the delete cookie path using the same ACS path.